Move the suggestions beside the glossary, not into the engine
src/sim is documented as the simulation engine: no React, no DOM, no
I/O, drivable from a script. These strings are none of that. Nothing in
the engine reads them and no script would ever want them; they are text
shown to a student, which is what src/content already holds for the
glossary.

Pure relocation. No behaviour change.
